requirements.txt
psutil>=5.9
Pillow>=10.0
py-cpuinfo>=9.0
httpx>=0.24

